2.1 Concluindo regras de negócio

+++ src/main/java/io/github/lorimedeiros/cadastro_autores_api/service/AutorService.java
@@ -1,7 +1,9 @@
 package io.github.lorimedeiros.cadastro_autores_api.service;
 
+import io.github.lorimedeiros.cadastro_autores_api.exceptions.OperacaoNaoPermitidaException;
 import io.github.lorimedeiros.cadastro_autores_api.model.Autor;
 import io.github.lorimedeiros.cadastro_autores_api.repository.AutorRepository;
+import io.github.lorimedeiros.cadastro_autores_api.repository.LivroRepository;
 import io.github.lorimedeiros.cadastro_autores_api.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -14,10 +16,12 @@ public class AutorService {
 
     private final AutorRepository repository;
     private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository, AutorValidator validator){
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository){
         this.repository = repository;
         this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
@@ -30,6 +34,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if (possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um autor que possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -50,6 +57,10 @@ public class AutorService {
         }
         validator.validar(autor);
         repository.save(autor);
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 
 }
